完成增删改查功能

+++ src/main/java/com/example/springboot/model/EmpUsers.java
@@ -2,13 +2,14 @@ package com.example.springboot.model;
 
 import lombok.Data;
 
+
 import java.util.Date;
 
 @Data
 public class EmpUsers extends EmpUsersKey {
     private String id;
 
-    private Integer order;
+    private Long order;
 
     private String department;
