+++ src/main/java/com/example/springboot/model/EmpUsersKey.java
@@ -6,6 +6,6 @@ import lombok.Data;
 public class EmpUsersKey {
     private String id;
 
-    private Integer order;
+    private Long order;
 
 }
